Add files via upload
The Hangman game is a classic word-guessing game that challenges players to identify a hidden word by guessing individual letters. The player has a limited number of incorrect guesses allowed, represented by "lives." With each incorrect guess, a part of the hangman is drawn. The objective is to guess the word before all lives are lost.